refactor: extract loadWorkspace into scripts/workspace.mjs
Both check:* IO shells import it; drops the argv main-guard from
check-catalog.mjs.
