+++ scripts/check-catalog.mjs
@@ -8,20 +8,11 @@
 // soft default; this check is the hard gate that keeps duplication from drifting
 // back in (e.g. via hand-edited package.json or a `--save-catalog`-less add).
 //
-// This file is the IO shell: it enumerates workspace members and reads the
-// workspace manifest via the pnpm SDK, then delegates all decisions to the pure,
-// unit-tested functions in ./check-catalog.core.mjs.
-//
-// Why the SDK (not the lockfile, not hand-parsed YAML): pnpm-lock.yaml has
-// already resolved `catalog:` refs to concrete versions, so it can't distinguish
-// a catalogued dep from an inline one — that lives only in package.json.
-// `findPackages` is pnpm's own workspace resolver, so we inherit its glob
-// handling, including that the standalone tutorials under examples/* are NOT
-// members (the glob is `examples`, not `examples/*`).
+// This file is the IO shell: it enumerates workspace members via
+// ./workspace.mjs, then delegates all decisions to the pure, unit-tested
+// functions in ./check-catalog.core.mjs.
 
-import { findPackages } from '@pnpm/fs.find-packages';
-import { readWorkspaceManifest } from '@pnpm/workspace.read-manifest';
-import { dirname, join, relative } from 'node:path';
+import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import {
@@ -30,26 +21,9 @@ import {
   findViolations,
   formatReport,
 } from './check-catalog.core.mjs';
+import { loadWorkspace } from './workspace.mjs';
 
 const ROOT = join(dirname(fileURLToPath(import.meta.url)), '..');
-
-/** Enumerate workspace members (incl. root) and the parsed workspace manifest. */
-export async function loadWorkspace(root) {
-  const workspaceManifest = await readWorkspaceManifest(root);
-  const projects = await findPackages(root, {
-    patterns: workspaceManifest?.packages,
-    includeRoot: true,
-  });
-  const members = projects.map((project) => {
-    const dir = relative(root, project.rootDir) || '<root>';
-    return {
-      name: project.manifest?.name ?? dir,
-      dir,
-      manifest: project.manifest,
-    };
-  });
-  return { members, workspaceManifest };
-}
 
 async function main() {
   const { members, workspaceManifest } = await loadWorkspace(ROOT);
@@ -62,10 +36,7 @@ async function main() {
   if (!ok) process.exitCode = 1;
 }
 
-// Guarded so check-pack.mjs can import loadWorkspace without running this check.
-if (process.argv[1] === fileURLToPath(import.meta.url)) {
-  main().catch((error) => {
-    console.error(error);
-    process.exitCode = 1;
-  });
-}
+main().catch((error) => {
+  console.error(error);
+  process.exitCode = 1;
+});
